+++ src/main/java/gameobject/character/enemy/Zako.java
@@ -9,7 +9,7 @@ import GameRegistrer.GameRegistrer;
 
 public class Zako extends Enemy {
 
-    public Zako(int x, int y, int size, Color color, String name, double bulletOffence, int bulletVelocityX, int bulletVelocityY, int shootInterval, int health) {
+    public Zako(int x, int y, int size, Color color, String name, double bulletOffence, double bulletVelocityX, double bulletVelocityY, int shootInterval, int health) {
         super(x, y, size, color, name, bulletOffence, bulletVelocityX, bulletVelocityY, shootInterval, health);
     }
